server/requirements.txt
websockets==12.0
mediapipe==0.10.9
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
